Restructuring into Python module layout
