+++ bin/run_control.sh
@@ -5,10 +5,11 @@
 # Define log file name for the data reading/writing program and start it from
 # the right location.
 ################################################################################
-WORKDIR="$(readlink -f "$(dirname "${0}")")"
+WORKDIR="$(readlink -f "$( dirname "$(dirname "${0}")")")"
+MODULEDIR="bocs_control"
 LOGDIR="${WORKDIR}/logs"
 LOGFILE="${LOGDIR}/control.log"
 PYTHON=$(command -v python3)
 
 cd "${WORKDIR}" || exit 1
-${PYTHON} "${WORKDIR}/control.py" > "${LOGFILE}" 2>&1
+${PYTHON} -m "${MODULEDIR}.control" > "${LOGFILE}" 2>&1
