+++ bocs_control/bocs_logger.py
@@ -1,0 +1,8 @@
+import logging
+
+logging.basicConfig(
+    format="[%(asctime)s.%(msecs)03d] %(levelname)s: %(filename)s:%(lineno)d %(message)s",
+    datefmt="%Y-%m-%d %H:%M:%S",
+    level=logging.INFO,
+    style="%",
+)
