+++ etc/systemd/system/bocs_control.service
@@ -5,7 +5,7 @@ Description=BOCS Data Collector
 WorkingDirectory=/home/pi/bocs_control
 User=pi
 Group=pi
-ExecStart=/bin/bash /home/pi/bocs_control/run_control.sh
+ExecStart=/bin/bash /home/pi/bocs_control/bin/run_control.sh
 Type=simple
 Restart=always
